requirements.txt
aiogram
openai
google-api-python-client
google-auth
python-dateutil
dateparser
python-dotenv